Move shared utilities into backend/shared
Clears the root folder of the backend a bit to clarify which parts are entry points.

This follows from these same files being in packages/shared/ since they're
shared by multiple renderers.
