+++ packages/react-devtools-shared/src/backend/shared/DevToolsConsolePatching.js
